Feature ETP-3662: extract CloneButton to shared to fix Sonar duplication
